Rettet typen på parameter capacity til function clear fra int -> intptr_t i klassen ByteArray Tilføjet funktioner til get/set af __int64/unsigned __int64 i Registry Afledte klasser BitSet8/16/32/64 er gjort til rigtige klasser (ike typedef) så de kan genkendes af debugger/natvis, samt tilføjet klasserne tilnatvis

+++ Lib/Include/ByteArray.h
@@ -32,7 +32,7 @@ public:
     return append(byte);
   }
 
-  ByteArray &clear(int newCapacity = 0);                   // return this, if newCapcity == -1 it will not be changed
+  ByteArray &clear(intptr_t newCapacity = 0);                   // return this, if newCapcity == -1 it will not be changed
   ByteArray &append( const ByteArray &d);
   ByteArray &append( const BYTE *data, size_t size);
   inline ByteArray &append(BYTE byte) {
